Clean repo structure. Rename and move to  'app/app.py' to prepare for deployment
