+++ app/recommendations/CourseRecommender.py
@@ -1,22 +1,27 @@
 import os
-from recommendation.models.base import BaseRecommender
+from app.recommendations.base import BaseRecommender
 from sklearn.feature_extraction.text import TfidfVectorizer
 from sklearn.metrics.pairwise import cosine_similarity
 import pandas as pd
 import numpy as np
 from sklearn.metrics import ndcg_score
+from sklearn.feature_extraction.text import CountVectorizer
+from nltk.corpus import stopwords
+from nltk.stem import WordNetLemmatizer
+import re
 
-class CourseRecommenderCosine(BaseRecommender):
+class CourseRecommender2(BaseRecommender):
     def __init__(self):
         super().__init__("course_recommender")
         self.vectorizer = None
         self.tfidf_matrix = None
         self.cosine_sim = None
+        self.stop_words = set(stopwords.words('english'))
+        self.lemmatizer = WordNetLemmatizer()
 
     def load_data(self):
         """
         Load and preprocess the training data.
-
         """
         # Load the training data (e.g., course descriptions)
         data_path = os.path.join("input_data", "kaggle_filtered_courses.csv")
@@ -25,8 +30,12 @@ class CourseRecommenderCosine(BaseRecommender):
         # Preprocess the data (e.g., normalize text, combine features)
         self.data['Course Description'] = self.data['Course Description'].str.lower()
         self.data['About'] = self.data['About'].str.lower()
+        # Ensure 'Difficulty Level' column exists and is of string type
+        if 'Difficulty Level' not in self.data.columns:
+            self.data['Difficulty Level'] = 'Unknown'  # Or some other default value
+        self.data['Difficulty Level'] = self.data['Difficulty Level'].astype(str).str.lower()
         self.data['combined_features'] = (
-            self.data['Name'] + ' ' + self.data['About'] + ' ' + self.data['Course Description']
+            self.data['Name'] + ' ' + self.data['About'] + ' ' + self.data['Course Description'] + ' ' + self.data['Difficulty Level']
         )
 
     def load_test_data(self):
@@ -49,47 +58,18 @@ class CourseRecommenderCosine(BaseRecommender):
             ]
         })
 
-    def evaluate(self, top_k=5):
-        """
-        Evaluate the model using its own test data.
-        """
-        precision_scores = []
-        recall_scores = []
-        ndcg_scores = []
-        
-        for _, row in self.test_data.iterrows():
-            query = row['query']
-            ground_truth = row['ground_truth']
-            
-            # Get recommendations
-            recommendations = self.predict(query, top_k, print_output=False) # Disable printing in evaluate
-            recommended_indices = recommendations.index.tolist()
-            
-            # Compute precision@k and recall@k
-            relevant = set(ground_truth)
-            retrieved = set(recommended_indices)
-            precision = len(relevant.intersection(retrieved)) / top_k
-            recall = len(relevant.intersection(retrieved)) / len(relevant) if len(relevant) > 0 else 0
-            
-            # Compute NDCG@k
-            relevance_scores = [1 if idx in ground_truth else 0 for idx in recommended_indices]
-            ndcg = ndcg_score([relevance_scores], [relevance_scores], k=top_k)
-            
-            precision_scores.append(precision)
-            recall_scores.append(recall)
-            ndcg_scores.append(ndcg)
-        
-        return {
-            'precisionk': np.mean(precision_scores),
-            'recallk': np.mean(recall_scores),
-            'ndcgk': np.mean(ndcg_scores)
-        }
-    
+    def preprocess_text(self, text):
+        text = re.sub(r'[^\w\s]', '', text)
+        text = ' '.join([word for word in text.split() if word not in self.stop_words])
+        text = ' '.join([self.lemmatizer.lemmatize(word) for word in text.split()])
+        return text
+
     def train(self):
         """
         Train the TF-IDF vectorizer and compute the cosine similarity matrix.
         """
-        self.vectorizer = TfidfVectorizer()
+        self.data['combined_features'] = self.data['combined_features'].apply(self.preprocess_text)
+        self.vectorizer = TfidfVectorizer(ngram_range=(1, 2), min_df=5) # Use n-grams
         self.tfidf_matrix = self.vectorizer.fit_transform(self.data['combined_features'])
         self.cosine_sim = cosine_similarity(self.tfidf_matrix, self.tfidf_matrix)
         self.is_trained = True
@@ -101,13 +81,13 @@ class CourseRecommenderCosine(BaseRecommender):
         Args:
             user_input (str): The user's input query.
             top_k (int): Number of recommendations to return.
-            print_output (bool): Whether to print the recommendations.
         
         Returns:
             pd.DataFrame: A DataFrame containing the recommended courses.
         """
         # Normalize the user input
         user_input = user_input.lower()
+        user_input = self.preprocess_text(user_input)
         
         # Vectorize the user input
         user_tfidf = self.vectorizer.transform([user_input])
@@ -123,6 +103,6 @@ class CourseRecommenderCosine(BaseRecommender):
         recommendations['Category'] = recommendations['Category'].astype(str)
         
         if print_output:
-            print(recommendations[['Name', 'University', 'Link', 'Category']])
+            print(recommendations[['Name', 'University', 'Link', 'Category', 'Difficulty Level']])
         
         return recommendations
